Fixed bug: only showing number of bases needed in `initField` (`player_count`)

+++ lib/game/Game.hpp
@@ -1,7 +1,7 @@
 
 class Game {
    public:
-	Game(TFT_Display* tft, LED_CONTROLLER* led, short player_count = 4);
+	Game(TFT_Display* tft, LED_CONTROLLER* led, u_int8_t player_count = 4);
 	~Game();
 
 	/**
